refactor: 식에서 숫자와 연산자를 분리하는 기능,연산자와 피연산자가 상호작용하는 기능 분리
