+++ src/main/kotlin/calculator/StringCalculator.kt
@@ -1,8 +1,6 @@
 package calculator
 
-private const val EXPRESSION_DELIMITER = " "
-private const val EVEN_AND_ODD_STANDARD = 2
-private const val NUMBER_POSITION = 0
+import calculator.util.ExpressionSeparator
 
 class StringCalculator(
     private val expression: String,
@@ -12,18 +10,8 @@ class StringCalculator(
     }
 
     fun calculate(): Int {
-        val numbersAndOperators = expression.split(EXPRESSION_DELIMITER)
-            .withIndex()
-            .partition { it.index % EVEN_AND_ODD_STANDARD == NUMBER_POSITION }
+        val (numbers, operators) = ExpressionSeparator.separate(expression)
 
-        val numbers = numbersAndOperators.first.map { it.value.toInt() }.toMutableList()
-        val operators = numbersAndOperators.second.map { Operator.of(it.value) }
-
-        var sum = numbers.removeFirst()
-        operators.forEach { operator ->
-            sum = operator.operate(sum, numbers.removeFirst())
-        }
-
-        return sum
+        return operators.operate(numbers.toMutableList())
     }
 }
